# src/main/resources/api.yml
openapi: 3.0.3
info:
  title: XYZ Bank - Account Service API
  description: API para gestionar cuentas bancarias de los clientes en el banco XYZ
  version: 1.0.0
servers:
  - url: http://localhost:8081/api
    description: Servidor local de desarrollo
paths:
  /cuentas:
    post:
      summary: Crear una nueva cuenta bancaria para un cliente
      operationId: createAccount
      requestBody:
        description: Datos de la cuenta a crear
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        '201':
          description: Cuenta creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDTO'
        '400':
          description: >-
            Entrada incorrecta o saldo inicial insuficiente (debe ser mayor que
            0)
        '409':
          description: El cliente ya tiene una cuenta bancaria con ese tipo
        '500':
          description: Error inesperado
    get:
      summary: Listar todas las cuentas bancarias
      operationId: listAccounts
      responses:
        '200':
          description: Lista de cuentas bancarias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountDTO'
        '404':
          description: No se encontraron cuentas
  /cuentas/{id}:
    get:
      summary: Obtener detalles de una cuenta bancaria por ID
      operationId: getAccountById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalles de la cuenta bancaria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDTO'
        '400':
          description: Entrada incorrecta
        '404':
          description: Cuenta no encontrada
        '500':
          description: Error inesperado
    delete:
      summary: Eliminar una cuenta bancaria
      operationId: deleteAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Cuenta eliminada exitosamente
        '400':
          description: Entrada incorrecta
        '404':
          description: Cuenta no encontrada
        '500':
          description: Error inesperado
  /cuentas/{cuentaId}/depositar:
    put:
      summary: Realizar un depósito en una cuenta bancaria
      operationId: depositToAccount
      parameters:
        - name: cuentaId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        description: Monto a depositar y cliente propietario de la cuenta
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  example: 100
                clienteId:
                  type: integer
                  example: 123
      responses:
        '200':
          description: Depósito exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDTO'
        '400':
          description: Entrada incorrecta o monto de depósito inválido
        '404':
          description: Cuenta o cliente no encontrado
        '409':
          description: El cliente no es el propietario de esta cuenta
        '500':
          description: Error inesperado
  /cuentas/{cuentaId}/retirar:
    put:
      summary: Realizar un retiro de una cuenta bancaria
      operationId: withdrawFromAccount
      parameters:
        - name: cuentaId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        description: Monto a retirar y cliente propietario de la cuenta
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  example: 50
                clienteId:
                  type: integer
                  example: 123
      responses:
        '200':
          description: Retiro exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDTO'
        '400':
          description: >-
            Retiro no permitido (saldo insuficiente para tipo de cuenta) o monto
            inválido
        '404':
          description: Cuenta o cliente no encontrado
        '409':
          description: El cliente no es el propietario de esta cuenta
        '500':
          description: Error inesperado
components:
  schemas:
    AccountDTO:
      type: object
      properties:
        id:
          type: integer
          example: 1
        numeroCuenta:
          type: string
          example: '000123456789'
        saldo:
          type: number
          example: 150
        tipoCuenta:
          type: string
          enum:
            - AHORROS
            - CORRIENTE
          example: AHORROS
        clienteId:
          type: integer
          example: 123
    AccountRequest:
      type: object
      properties:
        clienteId:
          type: integer
          example: 123
        saldoInicial:
          type: number
          minimum: 10
          example: 20
        tipoCuenta:
          type: string
          enum:
            - AHORROS
            - CORRIENTE
          example: AHORROS